Refactor name of service CourseSearch to CouserFinder
