+++ src/mooc/main/tv/codely/mooc/courses/application/update/CourseUpdater.java
@@ -2,7 +2,7 @@ package tv.codely.mooc.courses.application.update;
 
 import tv.codely.mooc.courses.domain.Course;
 import tv.codely.mooc.courses.domain.CourseRepository;
-import tv.codely.mooc.courses.domain.service.CourseSearch;
+import tv.codely.mooc.courses.domain.service.CourseFinder;
 import tv.codely.shared.domain.Service;
 
 import java.util.Optional;
@@ -10,15 +10,15 @@ import java.util.Optional;
 @Service
 public class CourseUpdater {
     private CourseRepository repository;
-    private CourseSearch search;
+    private CourseFinder finder;
 
-    public CourseUpdater (CourseRepository repository, CourseSearch search){
+    public CourseUpdater (CourseRepository repository, CourseFinder finder){
         this.repository = repository;
-        this.search = search;
+        this.finder = finder;
     }
 
     public void update(String id, String name) throws Exception{
-        Optional<Course> course = search.findOneById(id);
+        Optional<Course> course = finder.findOneById(id);
         if (course.isPresent()){
             repository.updateName(id, name);
         }
