+++ src/mooc/test/tv/codely/mooc/courses/domain/service/CourseFinderShould.java
@@ -1,22 +1,21 @@
-package tv.codely.mooc.courses.domain.service.search;
+package tv.codely.mooc.courses.domain.service;
 
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 import tv.codely.mooc.courses.domain.Course;
 import tv.codely.mooc.courses.domain.CourseRepository;
-import tv.codely.mooc.courses.domain.service.CourseSearch;
 
 import java.util.Optional;
 
 import static org.mockito.Mockito.*;
 
-final class CourseSearchShould {
+final class CourseFinderShould {
     @Test
     void get_a_valid_course() {
         CourseRepository repository = mock(CourseRepository.class);
-        CourseSearch search    = new CourseSearch(repository);
+        CourseFinder finder = new CourseFinder(repository);
 
-        search.last();
+        finder.last();
 
         verify(repository, atLeastOnce()).searchLast();
     }
@@ -24,11 +23,11 @@ final class CourseSearchShould {
     @Test
     void get_last_valid_course() {
         CourseRepository repository = mock(CourseRepository.class);
-        CourseSearch search    = new CourseSearch(repository);
+        CourseFinder finder = new CourseFinder(repository);
         Course course = new Course("id", "name", "duration");
         when(repository.searchLast()).thenReturn(Optional.of(course));
 
-        Optional<Course> courseFinded = search.last();
+        Optional<Course> courseFinded = finder.last();
 
         Assert.assertEquals(courseFinded.get().id(), course.id());
         Assert.assertEquals(courseFinded.get().name(), course.name());
